flask_certs/server/requirements.txt
pymongo==3.11.2
Flask==0.12.2
requests==2.25.1
pyopenssl==20.0.1
